+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version 1.0.2 stable (2013-02-??)
 
 ### Fixed
 - Added the missing labels
+- Fixed the incompatibility with Contao 3.0.3 and above (see #2)
 
 
 Version 1.0.1 stable (2013-02-12)